Revert useShare, then improve performance so e2e tests still pass
