+++ src/components/App.js
@@ -1,8 +1,9 @@
 // @flow
 
+import { useNavigation } from "@react-navigation/native";
 import RootDrawerNavigator from "navigation/rootDrawerNavigator";
 import type { Node } from "react";
-import React from "react";
+import React, { useCallback } from "react";
 import { log } from "sharedHelpers/logger";
 import {
   useCurrentUser,
@@ -18,6 +19,29 @@ import StartupService from "./StartupService";
 
 const logger = log.extend( "App" );
 
+type SharedItem = {
+  mimeType: string,
+  data: string | Array<string>
+};
+
+const handleShare = ( navigation, item: ?SharedItem ) => {
+  if ( !item ) {
+    // user hasn't shared any items
+    return;
+  }
+
+  const { mimeType, data } = item;
+
+  if ( !mimeType && !data ) {
+    // user hasn't shared any images
+    return;
+  }
+
+  // show user a loading animation screen (like PhotoLibrary)
+  // while observations are created
+  navigation?.navigate( "PhotoSharing", { item } );
+};
+
 type Props = {
   // $FlowIgnore
   children?: unknown,
@@ -26,6 +50,7 @@ type Props = {
 // this children prop is here for the sake of testing with jest
 // normally we would never do this in code
 const App = ( { children }: Props ): Node => {
+  const navigation = useNavigation( );
   const { loadTime } = usePerformance( {
     screenName: "App"
   } );
@@ -33,10 +58,17 @@ const App = ( { children }: Props ): Node => {
     logger.info( loadTime );
   }
 
+  // attempting to make sure that navigation is only called once
+  // for performance reasons
+  const onShare = useCallback(
+    item => handleShare( navigation, item ),
+    [navigation]
+  );
+
   const currentUser = useCurrentUser( );
 
   useLinking( currentUser );
-  useShare( );
+  useShare( onShare );
 
   // this children prop is here for the sake of testing with jest
   // normally we would never do this in code
